// vue/src/store/generated/threefold/threefold_hub/threefold.threefoldhub.validators/module/types/validators/params.d.ts
import { Writer, Reader } from "protobufjs/minimal";
export declare const protobufPackage = "threefold.threefoldhub.validators";
/** Params defines the parameters for the module. */
export interface Params {
    /** max_validators is the maximum number of validators. */
    maxValidators: number;
    /** historical_entries is the number of historical entries to persist. */
    historicalEntries: number;
}
export declare const Params: {
    encode(message: Params, writer?: Writer): Writer;
    decode(input: Reader | Uint8Array, length?: number): Params;
    fromJSON(object: any): Params;
    toJSON(message: Params): unknown;
    fromPartial(object: DeepPartial<Params>): Params;
};
declare type Builtin = Date | Function | Uint8Array | string | number | undefined;
export declare type DeepPartial<T> = T extends Builtin ? T : T extends Array<infer U> ? Array<DeepPartial<U>> : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>> : T extends {} ? {
    [K in keyof T]?: DeepPartial<T[K]>;
} : Partial<T>;
export {};